terms and policy

+++ models/whatsapp_config.go
@@ -11,7 +11,7 @@ const (
 // One row per user (multi-tenant).
 type WhatsAppConfig struct {
 	ID            int64      `gorm:"primary_key;AUTO_INCREMENT" json:"id"`
-	UserID        uint       `gorm:"not null;unique_index" json:"user_id"`
+	UserID        int64      `gorm:"not null;unique_index" json:"user_id"`
 	PhoneNumberID string     `gorm:"column:phone_number_id;not null" json:"phone_number_id"`
 	AccessToken   string     `gorm:"column:access_token;not null" json:"access_token"`
 	ApiVersion    string     `gorm:"column:api_version;not null;default:'v24.0'" json:"api_version"`
